Update package.json
chore(build): Add Pagefind script to package.json

This commit adds a build script to the package.json file. This script executes Pagefind, which will scan the site's HTML files and generate a static search index.

This is a key step in integrating a client-side search feature into the site's Netlify build process.